Bugfix: make it compile on 32-bit platforms.

+++ danmaku/dmpkg/raw.go
@@ -6,7 +6,7 @@ import (
 	"math"
 )
 
-const kMaxBodyLength = math.MaxUint32 - kHeaderLength
+const kMaxBodyLength = math.MaxUint32 - uint64(kHeaderLength)
 
 // NewPlainExchange creates a new exchange with raw body specified.
 // body: a struct or a raw string
